Move darknet and doodlenet classes to the imageClassifier folder
