+++ src/ImageClassifier/DOODLENET_CLASSES.js
